backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pandas
numpy
python-dotenv
requests
